Create FizzBuzz.js
This program uses a 'for' loop to iterate through the numbers from 1 to 100. The 'if-else' statements check if the current number is a multiple of both 3 and 5 (in which case "FizzBuzz" is printed), a multiple of 3 (in which case "Fizz" is printed), a multiple of 5 (in which case "Buzz" is printed), or none of the above (in which case the number itself is printed).